Add function safe_realloc
Replace all the malloc and realloc in the project with safe_realloc.

The safe_realloc is another implement of reallocarr(3). It checks the
arguments "number" and "size". If number or size is less than 0, or if
the product of them is larger than max size, the function will set errno
and return non-zero.

Reference: REALLOCARR(3)